% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{clean_results}
\alias{clean_results}
\title{Clean one fixture of scraped results}
\usage{
clean_results(raw_results)
}
\arguments{
\item{raw_results}{}
}
\value{

}
\description{
Clean one fixture of scraped results
}
