% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{concat}
\alias{concat}
\title{Concatenate strings together with a comma-separated list}
\usage{
concat(strings)
}
\arguments{
\item{strings}{a list of strings}
}
\value{

}
\description{
Concatenate strings together with a comma-separated list
}
