% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc.R
\name{ctmc_build_rate_matrix}
\alias{ctmc_build_rate_matrix}
\title{Title}
\usage{
ctmc_build_rate_matrix(lambda, n_states = 3, n_absorbing = 1)
}
\arguments{
\item{lambda}{}

\item{n_states}{}

\item{n_absorbing}{}
}
\value{

}
\description{
Title
}
