% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc.R
\name{ctmc_log_likelihood}
\alias{ctmc_log_likelihood}
\title{Title}
\usage{
ctmc_log_likelihood(rate_matrix, tau, observed_transitions)
}
\arguments{
\item{rate_matrix}{transition rate-matrix}

\item{tau}{a vector containing total time in each state}

\item{observed_transitions}{a matrix with all observed transitions}
}
\value{

}
\description{
Title
}
