% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmc.R
\name{ctmc_sim_exact}
\alias{ctmc_sim_exact}
\title{Simulate from a continuous time markov chain}
\usage{
ctmc_sim_exact(x, Q, n)
}
\arguments{
\item{x}{an integer, the initial state}

\item{Q}{a transition rate matrix}

\item{n}{the number of transitions to simulates}
}
\value{

}
\description{
Simulate from a continuous time markov chain
}
