% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_averaging.R
\name{da_step}
\alias{da_step}
\title{Title}
\usage{
da_step(
  log_posterior,
  gradient,
  position,
  initial_step_size,
  n_steps,
  iteration,
  warmup_iterations,
  d
)
}
\arguments{
\item{log_posterior}{}

\item{gradient}{}

\item{position}{}

\item{initial_step_size}{}

\item{n_steps}{}

\item{iteration}{}

\item{warmup_iterations}{}

\item{d}{}
}
\value{

}
\description{
Title
}
