% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{density_plot}
\alias{density_plot}
\title{Plot a density plot from MCMC output}
\usage{
density_plot(chains)
}
\arguments{
\item{chains}{}
}
\value{

}
\description{
Plot a density plot from MCMC output
}
