% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_center_y}
\alias{dlm_center_y}
\title{Helper function for Kalman Filter and Gibbs Sampling}
\usage{
dlm_center_y(ys, theta, mod)
}
\arguments{
\item{ys}{p x n matrix of observation}

\item{theta}{d x n matrix of sampled latent state}

\item{mod}{the dlm model}
}
\value{

}
\description{
Centre the observations
}
