% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_ffbs}
\alias{dlm_ffbs}
\title{Perform Forward Filtering Backward Sampling}
\usage{
dlm_ffbs(ys, mod)
}
\arguments{
\item{ys}{p x n matrix of observation}

\item{mod}{the dlm model}
}
\value{

}
\description{
Perform Forward Filtering Backward Sampling
}
