% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_gibbs}
\alias{dlm_gibbs}
\title{Gibbs sampling DLM}
\usage{
dlm_gibbs(ys, mod, shape_v, rate_v, shape_w, rate_w, iters, progress = FALSE)
}
\arguments{
\item{iters}{}
}
\value{

}
\description{
Sample diagonal system and noise covariance matrices using a Gibbs sampler
}
