% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm_gibbs_dgamma}
\alias{dlm_gibbs_dgamma}
\title{Perform Gibbs Sampling for the DLM}
\usage{
dlm_gibbs_dgamma(y, mod, theta, shape_v, rate_v, shape_w, rate_w, m)
}
\arguments{
\item{mod}{a function from theta = c(v, w) => KFAS model object}

\item{m}{}
}
\value{

}
\description{
Perform Gibbs Sampling for the DLM
}
