% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_gibbs_parallel}
\alias{dlm_gibbs_parallel}
\title{Gibbs sample DLM in parallel}
\usage{
dlm_gibbs_parallel(
  ys,
  mod,
  shape_v,
  rate_v,
  shape_w,
  rate_w,
  iters,
  chains = 4
)
}
\arguments{
\item{chains}{}
}
\value{

}
\description{
Run 4 chains in parallel for DLM with diagonal state and processes noise
covariance matrices
}
