% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_gibbs_wishart}
\alias{dlm_gibbs_wishart}
\title{Title}
\usage{
dlm_gibbs_wishart(ys, mod, nu_v, psi_v, nu_w, psi_w, iters, progress = FALSE)
}
\arguments{
\item{ys}{}

\item{mod}{}

\item{nu_v}{}

\item{psi_v}{}

\item{nu_w}{}

\item{psi_w}{}

\item{iters}{}

\item{progress}{}
}
\value{

}
\description{
Title
}
