% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_non_missing}
\alias{dlm_non_missing}
\title{Helper function for Kalman Filter}
\usage{
dlm_non_missing(y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Sum the non-missing observations in each column of the observation matrix
}
