% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_sample_v_wishart}
\alias{dlm_sample_v_wishart}
\title{Sample the observation noise matrix}
\usage{
dlm_sample_v_wishart(y, theta, mod, shape_v, rate_v)
}
\arguments{
\item{theta}{d x n matrix of the sampled latent state}

\item{mod}{the dlm model}

\item{shape_v}{the prior shape of the IG distribution}

\item{rate_v}{the prior rate of the IG distribution}

\item{ys}{p x n matrix of observation}
}
\value{

}
\description{
Sample the observation noise matrix from a d-Inverse Gamma distribution
}
