% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_sample_w}
\alias{dlm_sample_w}
\title{Sample the system noise matrix}
\usage{
dlm_sample_w(theta, mod, shape_w, rate_w)
}
\arguments{
\item{theta}{a d x n matrix containing the sampled latent state}

\item{mod}{the dlm model}

\item{rate_w}{}
}
\value{

}
\description{
Sample the system noise matrix from a d-inverse Gamma distribution
}
