% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_summarise_filtered}
\alias{dlm_summarise_filtered}
\title{Calculate one-step forecast mean and variance for each observation}
\usage{
dlm_summarise_filtered(filtered)
}
\arguments{
\item{filtered}{output from dlm_kalman_filter}
}
\value{

}
\description{
Calculate one-step forecast mean and variance for each observation
}
