% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{dlm_to_kfas}
\alias{dlm_to_kfas}
\title{Transform a dlm to a KFAS State Space Model}
\usage{
dlm_to_kfas(dlm_model, ys, theta)
}
\arguments{
\item{dlm_model}{a model created using the dlm package}

\item{ys}{a matrix containing observations}

\item{theta}{a vector of parameters containing the diagonal values of v and w}
}
\value{

}
\description{
Transform a dlm to a KFAS State Space Model
}
