% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Effective Sample Size Particle Filter}
\usage{
effective_sample_size(weights)
}
\arguments{
\item{weights}{}
}
\value{

}
\description{
Calculate the effective sample size of the particle cloud.
This can be used as a condition for resampling.
}
