% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{effective_size}
\alias{effective_size}
\title{Calculate Effective Sample Size}
\usage{
effective_size(values)
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Calculate Effective Sample Size
}
