% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehmc.R
\name{ehmc}
\alias{ehmc}
\title{Hamiltonian Monte Carlo}
\usage{
ehmc(log_posterior, gradient, n_steps, init_parameters, iters, chains = 2)
}
\arguments{
\item{log_posterior}{a function for the unnormalised log-posterior
from parameters -> log-likelihood}

\item{gradient}{the gradient of the log_posterior}

\item{n_steps}{the number of leapfrog steps}

\item{init_parameters}{a named vector of initial parameters}

\item{iters}{the number of iterations}

\item{chains}{the number of chains}

\item{step_size}{the step size of the leapfrog steps}
}
\value{

}
\description{
Run multiple HMC chains and return a dataframe of parameters
}
\details{
To run this in parallel call
\code{future::plan(future::multiprocess())} before this function
}
