% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehmc.R
\name{ehmc_helper}
\alias{ehmc_helper}
\title{Empirical Hamiltonian Monte Carlo}
\usage{
ehmc_helper(
  log_posterior,
  gradient,
  n_steps,
  init_parameters,
  iters,
  warmup_iters = iters/2,
  logging = TRUE
)
}
\arguments{
\item{log_posterior}{a function for the unnormalised log-posterior
from parameters -> log-likelihood}

\item{gradient}{the gradient of the log_posterior}

\item{n_steps}{the initial number of leapfrog steps}

\item{init_parameters}{a named vector containing the initial parameters}

\item{iters}{the number of iterations}
}
\value{
a matrix of iterations from the HMC algorithm representing draws
from the posterior distribution
}
\description{
Perform Empirical Hamiltonian Monte Carlo
}
