% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{extract_date}
\alias{extract_date}
\title{Get the date of a fixture}
\usage{
extract_date(results_home)
}
\arguments{
\item{results_home}{the url of a fixture}
}
\value{

}
\description{
Extract the date from an h1 or h2 title on the results page
}
