% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_averaging.R
\name{find_reasonable_epsilon}
\alias{find_reasonable_epsilon}
\title{Title}
\usage{
find_reasonable_epsilon(position, momentum, log_posterior, gradient, d)
}
\arguments{
\item{position}{}

\item{momentum}{}

\item{log_posterior}{}

\item{gradient}{}

\item{d}{}
}
\value{

}
\description{
Title
}
