% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{fitted_kalman_filter}
\alias{fitted_kalman_filter}
\title{Title}
\usage{
fitted_kalman_filter(ys, ws, vs, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Perform the Kalman filter using draws from the posterior distribution
}
