% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{format_course_name}
\alias{format_course_name}
\title{Title}
\usage{
format_course_name(parsed_courses, possible_courses)
}
\arguments{
\item{parsed_courses}{}

\item{possible_courses}{}
}
\value{

}
\description{
Title
}
