% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{get_course_list}
\alias{get_course_list}
\title{Get a list of courses for a given year}
\usage{
get_course_list(year)
}
\arguments{
\item{year}{}
}
\value{

}
\description{
Get a list of courses for a given year
}
