% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{get_division_results}
\alias{get_division_results}
\title{Get Harrier League Results}
\usage{
get_division_results(raw_results, counters = 4)
}
\arguments{
\item{raw_results}{the finishing positions of the runners}

\item{counters}{the total number of counters to use for team results}
}
\value{

}
\description{
This function calculates the position of a club in their respective division
by summing the positions of the first n = \code{counters} and sorting the sum
from smallest to largest.
}
