% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{get_harrier_league_results}
\alias{get_harrier_league_results}
\title{Fetch a table containing the harrier league results}
\usage{
get_harrier_league_results(url)
}
\arguments{
\item{url}{the url of a results page}
}
\value{

}
\description{
Fetch a table containing the harrier league results
}
