% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{get_results_home}
\alias{get_results_home}
\title{Get the URL of a single fixture}
\usage{
get_results_home(courses, result_type = "SenM")
}
\arguments{
\item{courses}{}

\item{result_type}{}
}
\value{

}
\description{
Get the URL of a single fixture
}
