% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc}
\alias{hmc}
\title{Hamiltonian Monte Carlo}
\usage{
hmc(
  log_posterior,
  gradient,
  step_size,
  num_steps,
  theta_0,
  iters,
  parameter_names
)
}
\arguments{
\item{parameter_names}{}
}
\value{

}
\description{
Perform HMC and return a dataframe of parameters
}
