% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_averaging.R
\name{hmc_da_helper}
\alias{hmc_da_helper}
\title{Hamiltonian Monte Carlo}
\usage{
hmc_da_helper(
  log_posterior,
  gradient,
  n_steps,
  init_parameters,
  iters,
  warmup_iters = iters/2,
  logging = TRUE
)
}
\arguments{
\item{log_posterior}{a function for the unnormalised log-posterior
from parameters -> log-likelihood}

\item{gradient}{the gradient of the log_posterior}

\item{n_steps}{the number of leapfrog steps}

\item{init_parameters}{a named vector containing the initial parameters}

\item{iters}{the number of iterations}

\item{warmup_iters}{the total number of warmup iterations to use for the adaptation phase}

\item{logging}{enable logging}
}
\value{
a matrix of iterations from the HMC algorithm representing draws
from the posterior distribution
}
\description{
Perform Hamiltonian Monte Carlo with dual averaging
}
