% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_df}
\alias{hmc_df}
\title{Hamiltonian Monte Carlo}
\usage{
hmc_df(log_posterior, gradient, step_size, l, theta_0, m, parameter_names)
}
\arguments{
\item{parameter_names}{}
}
\value{

}
\description{
Perform HMC and return a dataframe
}
