% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_helper}
\alias{hmc_helper}
\title{Hamiltonian Monte Carlo}
\usage{
hmc_helper(log_posterior, gradient, step_size, n_steps, init_parameters, iters)
}
\arguments{
\item{log_posterior}{a function for the unnormalised log-posterior
from parameters -> log-likelihood}

\item{gradient}{the gradient of the log_posterior}

\item{step_size}{the step size of the leapfrog steps}

\item{n_steps}{the number of leapfrog steps}

\item{init_parameters}{a named vector containing the initial parameters}

\item{iters}{the number of iterations}
}
\value{
a matrix of iterations from the HMC algorithm representing draws
from the posterior distribution
}
\description{
Perform Hamiltonian Monte Carlo
}
