% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_leapfrog_step}
\alias{hmc_leapfrog_step}
\title{Leapfrog Step}
\usage{
hmc_leapfrog_step(gradient, step_size, position, momentum, d)
}
\arguments{
\item{momentum}{}
}
\value{

}
\description{
Leapfrog Step
}
