% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_leapfrogs}
\alias{hmc_leapfrogs}
\title{Perform l Leapfrog steps}
\usage{
hmc_leapfrogs(gradient, step_size, l, position, momentum, d)
}
\arguments{
\item{momentum}{}
}
\value{

}
\description{
Perform l Leapfrog steps
}
