% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_log_acceptance}
\alias{hmc_log_acceptance}
\title{Log Acceptance for the HMC algorithm}
\usage{
hmc_log_acceptance(
  prop_position,
  prop_momentum,
  position,
  momentum,
  log_posterior
)
}
\arguments{
\item{log_posterior}{}
}
\value{

}
\description{
Log Acceptance for the HMC algorithm
}
