% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_mat}
\alias{hmc_mat}
\title{Hamiltonian Monte Carlo}
\usage{
hmc_mat(log_posterior, gradient, step_size, num_steps, theta_0, iters)
}
\arguments{
\item{iters}{}
}
\value{

}
\description{
Perform Hamiltonian Monte Carlo
}
