% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc_step}
\alias{hmc_step}
\title{HMC Step}
\usage{
hmc_step(log_posterior, gradient, step_size, n_steps, position, d)
}
\arguments{
\item{log_posterior}{a function for the unnormalised log-posterior
from parameters -> log-likelihood}

\item{gradient}{the gradient of the log_posterior}

\item{step_size}{the step size of the leapfrog steps}

\item{n_steps}{the number of leapfrog steps}

\item{position}{the current value of the parameter vector}

\item{d}{the dimension of the parameter vector}
}
\value{

}
\description{
HMC Step
}
