% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_backward}
\alias{hmm_backward}
\title{Title}
\usage{
hmm_backward(ys, transition_matrix, observation)
}
\arguments{
\item{ys}{}

\item{transition_matrix}{}

\item{observation}{}
}
\value{

}
\description{
Title
}
