% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_backward_step}
\alias{hmm_backward_step}
\title{HMM Backward Smoothing}
\usage{
hmm_backward_step(beta, y, transition_matrix, observation)
}
\arguments{
\item{beta}{}

\item{y}{}

\item{transition_matrix}{the transition matrix}

\item{observation}{the observation distribution}
}
\value{

}
\description{
HMM Backward Smoothing
}
