% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_forward_backward}
\alias{hmm_forward_backward}
\title{Perform smoothing using the forward-backward algorithm}
\usage{
hmm_forward_backward(ys, x0, transition_matrix, observation)
}
\arguments{
\item{ys}{The observed values}

\item{x0}{the distribution of the initial state}

\item{transition_matrix}{the transition matrix}

\item{observation}{the observation distribution}
}
\value{

}
\description{
Perform smoothing using the forward-backward algorithm
}
