% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_forward_step}
\alias{hmm_forward_step}
\title{Forward step for HMM}
\usage{
hmm_forward_step(alpha, y, transition_matrix, observation)
}
\arguments{
\item{alpha}{The previous posterior distribution over the discrete states of the HMM}

\item{transition_matrix}{the transition matrix}

\item{observation}{the observation distribution}

\item{x0}{the initial state distribution}
}
\value{

}
\description{
A single forward step of the forward filtering algorithm for a hidden Markov model
}
