% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_log_likelihood}
\alias{hmm_log_likelihood}
\title{Marginal log-likelihood HMM}
\usage{
hmm_log_likelihood(ys, x0, observation, transition_matrix)
}
\arguments{
\item{ys}{a sequence of observations}

\item{x0}{the initial state distribution}

\item{observation}{the observation distribution}

\item{transition_matrix}{the transition matrix}
}
\value{

}
\description{
Calculate the marginal log-likelihood for an HMM using the forward filtering algorithm
}
