% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_simulate}
\alias{hmm_simulate}
\title{Simulate from a hidden Markov model}
\usage{
hmm_simulate(n, x0, transition_matrix, sim_observation)
}
\arguments{
\item{n}{the number of observations}

\item{x0}{the initial distribution}

\item{transition_matrix}{the transition matrix}

\item{sim_observation}{the observation distribution p(y_t|x_t)}
}
\value{

}
\description{
Simulate from a hidden Markov model
}
