% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{iterpolate_site}
\alias{iterpolate_site}
\title{Perform Kalman Smoother on a single site}
\usage{
iterpolate_site(site, model)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Perform Kalman Smoother on a single site
}
