% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{leapfrogs}
\alias{leapfrogs}
\title{Perform l Leapfrog steps}
\usage{
leapfrogs(gradient, step_size, l, position, momentum, d)
}
\arguments{
\item{momentum}{}
}
\value{

}
\description{
Perform l Leapfrog steps
}
