% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{ll_step}
\alias{ll_step}
\title{Title}
\usage{
ll_step(state, y, observation, transition_matrix)
}
\arguments{
\item{state}{}

\item{y}{}

\item{observation}{}

\item{transition_matrix}{}
}
\value{

}
\description{
Title
}
