% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{log_acceptance}
\alias{log_acceptance}
\title{Log Acceptance for the HMC algorithm}
\usage{
log_acceptance(prop_position, prop_momentum, position, momentum, log_posterior)
}
\arguments{
\item{log_posterior}{}
}
\value{

}
\description{
Log Acceptance for the HMC algorithm
}
