% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehmc.R
\name{longest_batch}
\alias{longest_batch}
\title{Title}
\usage{
longest_batch(gradient, position, momentum, step_size, n_steps, d)
}
\arguments{
\item{position}{}

\item{momentum}{}

\item{step_size}{}

\item{n_steps}{}

\item{d}{}
}
\value{

}
\description{
Find the longest batch for a singe position
}
