% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metropolis_helper}
\alias{metropolis_helper}
\title{Metropolis Helper}
\usage{
metropolis_helper(theta, log_posterior, proposal, m)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
This implements the metropolis algorithm given an initial parameter vector,
the un-normalised log-posterior and a symmetric proposal distribution
}
