% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metropolis_step}
\alias{metropolis_step}
\title{Metropolis Step}
\usage{
metropolis_step(theta, log_posterior, proposal)
}
\arguments{
\item{proposal}{}
}
\value{

}
\description{
This implements a single step of the Metropolis Algorithm
}
