% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{non_missing}
\alias{non_missing}
\title{Title}
\usage{
non_missing(y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Sum the non-missing observations in each column of the observation matrix
}
