% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{normalise}
\alias{normalise}
\title{Normalise such that the values of the vector sum to one}
\usage{
normalise(a)
}
\arguments{
\item{a}{a vector of numbers}
}
\value{

}
\description{
Normalise such that the values of the vector sum to one
}
