% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{not_all_na}
\alias{not_all_na}
\title{Check if some columns contain non-missing values}
\usage{
not_all_na(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Check if some columns contain non-missing values
}
