% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harrier_league.R
\name{parse_results_table}
\alias{parse_results_table}
\title{Get the raw HTML table from HTML}
\usage{
parse_results_table(html)
}
\arguments{
\item{html}{a string containing html for the page}
}
\value{

}
\description{
Get the raw HTML table from HTML
}
