% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R, R/particle_filter.R
\name{plot.filtered}
\alias{plot.filtered}
\title{Plot the one-step forecast}
\usage{
\method{plot}{filtered}(x)

\method{plot}{filtered}(x)
}
\arguments{
\item{x}{an object of class "filtered" returns from the function "particle_filter"}

\item{filtered}{output of dlm_kalman_filter of class "filtered"}
}
\value{



}
\description{
Generic S3 method to plot the one-step forecast mean f_t = F_t a_t and variance
Q_t = F_t r_t F_t^T + V_t against the observed values
}
