% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{plot_diagnostics_sim}
\alias{plot_diagnostics_sim}
\title{Plot MCMC diagnostics with known values}
\usage{
plot_diagnostics_sim(chains, actual_values)
}
\arguments{
\item{chains}{output from an MCMC sampler}

\item{actual_values}{dataframe containing the values used to simulate synthetic data}
}
\value{

}
\description{
Plot MCMC diagnostics with known values
}
