% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{prior_posterior}
\alias{prior_posterior}
\title{Plot the prior and posterior}
\usage{
prior_posterior(chain, parameters, prior, limits)
}
\arguments{
\item{chain}{}

\item{parameters}{}

\item{prior}{}

\item{limits}{}
}
\value{

}
\description{
Plot the prior and posterior
}
