% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{sample_v}
\alias{sample_v}
\title{Title}
\usage{
sample_v(y, theta, mod, shape_v, rate_v)
}
\arguments{
\item{y}{}

\item{theta}{}

\item{mod}{}

\item{shape_v}{}

\item{rate_v}{}
}
\value{

}
\description{
Title
}
