% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{sample_w}
\alias{sample_w}
\title{Title}
\usage{
sample_w(theta, mod, shape_w, rate_w)
}
\arguments{
\item{theta}{}

\item{mod}{}

\item{shape_w}{}

\item{rate_w}{}
}
\value{

}
\description{
Title
}
