% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_dlm.R
\name{spatial_dlm}
\alias{spatial_dlm}
\title{Title}
\usage{
spatial_dlm(
  ys,
  xs,
  mod,
  shape_v,
  rate_v,
  shape_w,
  rate_w,
  beta_mu,
  beta_sigma,
  iters,
  progress = TRUE
)
}
\arguments{
\item{ys}{}

\item{xs}{}

\item{mod}{}

\item{shape_v}{}

\item{rate_v}{}

\item{shape_w}{}

\item{rate_w}{}

\item{beta_mu}{}

\item{beta_sigma}{}

\item{iters}{}

\item{progress}{}
}
\value{

}
\description{
Title
}
