% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_dlm.R
\name{spatial_dlm_parallel}
\alias{spatial_dlm_parallel}
\title{Gibbs sample spatial DLM in parallel}
\usage{
spatial_dlm_parallel(
  ys,
  xs,
  mod,
  shape_v,
  rate_v,
  shape_w,
  rate_w,
  beta_mu,
  beta_sigma,
  iters = 10000,
  chains = 4
)
}
\arguments{
\item{chains}{}
}
\value{

}
\description{
Run 4 chains in parallel for DLM with diagonal state and processes noise
covariance matrices
}
