% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sample_dlm.R
\name{summary.filtered}
\alias{summary.filtered}
\title{Calculate one-step forecast mean and variance for each observation}
\usage{
\method{summary}{filtered}(filtered, interval = 0.9)
}
\arguments{
\item{filtered}{output from dlm_kalman_filter}

\item{interval}{the width of the probability interval}
}
\value{

}
\description{
S3 generic to calculate the one-step forecast mean f_t = F_t a_t and covariance
Q_t = F_t r_t F_t^T + V_t then collect the mean and marginal variance for each forecast
into a tibble.
}
\examples{
model <- dlmRandom(1, 1)
filtered <-
  dlm_kalman_filter(
    ys = t(model$y),
    f = model$mod$FF,
    g = model$mod$GG,
    v = model$mod$V,
    w = model$mod$W,
    m0 = model$mod$m0,
    c0 = model$mod$C0
  )
summary(filtered)
}
