% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{thin}
\alias{thin}
\title{Thin MCMC}
\usage{
thin(df, nth)
}
\arguments{
\item{df}{a dataframe containing one sample per from from an MCMC}

\item{nth}{the number to thin by}
}
\value{

}
\description{
Thin a dataframe by only keeping every nth row
}
