% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{traceplot}
\alias{traceplot}
\title{Plot a traceplot}
\usage{
traceplot(chains)
}
\arguments{
\item{paginate}{boolean indicating whether to paginate the plots}
}
\value{

}
\description{
Plot a traceplot
}
