% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_averaging.R
\name{update_step_size}
\alias{update_step_size}
\title{Title}
\usage{
update_step_size(
  iteration,
  accept_prob,
  mu,
  delta = 0.65,
  k = 0.75,
  gamma = 0.05,
  t0 = 10
)
}
\arguments{
\item{iteration}{the current iteration}

\item{accept_prob}{the acceptance probability from the previous time step}

\item{mu}{tuning parameter, default set to log(10 * eps0)}

\item{delta}{target acceptance probability, default 0.65}

\item{k}{tuning parameter, default set to 0.75}

\item{gamma}{tuning parameter, default set to 0.05}

\item{t0}{tuning parameter, default set to 10.0}

\item{hm0}{the previous state}

\item{logeps0}{the previous state}

\item{logepsbar0}{the previous state}
}
\value{

}
\description{
Update the value of epsilon during an adaptation phase
}
