% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateState}
\alias{updateState}
\title{Update the state using Joseph Form Update given the newly observed data.}
\usage{
updateState(ft, at, rt, predicted, predcov, y, v)
}
\description{
Update the state using Joseph Form Update given the newly observed data.
}
