% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_backward_step}
\alias{hmm_backward_step}
\title{Title}
\usage{
hmm_backward_step(beta, y, observation, transition_matrix)
}
\arguments{
\item{beta}{}

\item{y}{}

\item{observation}{}

\item{transition_matrix}{}
}
\value{

}
\description{
Title
}
