% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm_forward}
\alias{hmm_forward}
\title{Forward Filtering an HMM}
\usage{
hmm_forward(ys, x0, transition_matrix, observation)
}
\arguments{
\item{ys}{a sequence of observations}

\item{x0}{the initial state distribution}

\item{transition_matrix}{the transition matrix}

\item{observation}{the observation distribution}
}
\value{

}
\description{
Forward Filtering an HMM
}
