% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_ad.R
\name{lift_function}
\alias{lift_function}
\title{Lift a function to operate on bare numbers}
\usage{
lift_function(f)
}
\arguments{
\item{f}{a function operating on dual numbers}
}
\value{

}
\description{
Lift a function to operate on bare numbers
}
