% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metropolis}
\alias{metropolis}
\title{Metropolis Algorithm}
\usage{
metropolis(theta, log_posterior, proposal, m, chains = 1, parallel = FALSE)
}
\arguments{
\item{parallel}{}
}
\value{

}
\description{
Metropolis Algorithm
}
