% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mcmc.R
\name{plot_pdf}
\alias{plot_pdf}
\title{Plot a probability density function}
\usage{
plot_pdf(pdf, mode, scale, range = c(-10, 10), title)
}
\arguments{
\item{title}{}
}
\value{

}
\description{
Plot a probability density function
}
\examples{
plot_pdf()
}
