% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{sample_v}
\alias{sample_v}
\title{Sampling the observation noise matrix}
\usage{
sample_v(y, state, FF, shape_v, rate_v)
}
\arguments{
\item{rate_v}{}
}
\value{

}
\description{
Sampling the observation noise matrix
}
