% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_dlm.R
\name{spatial_dlm}
\alias{spatial_dlm}
\title{Title}
\usage{
spatial_dlm(
  ys,
  xs,
  mod,
  shape_v,
  rate_v,
  shape_w,
  rate_w,
  beta_mu,
  beta_sigma,
  iters,
  progress = TRUE
)
}
\arguments{
\item{progress}{}
}
\value{

}
\description{
Title
}
